+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 ### Changed
 - Minimum dart version bumped to 3.8
+- Internal implementation changed from AnimatedBuilder to ListenableBuilder
 
 ## 0.1.1 - 2024-11-26
 ### Fixed